Review Research
This is an exercise in the old guide that was not yet brought over.

It might have been intentional as this is a broad exercise that can also
be broken down into several other exercises.

That being said, there are times where a less-structured overview of all
of the existing research and context might make sense.

Opening this PR for discussion. If we want to keep this, should it make
its way onto one of the schedules?

If we don't want to keep there, where should it redirect to?