Group Critique
This is an exercise in the old guide that was not yet brought over.

It might have been intentional as this is an that can sometimes devolve
into too much of a discussion.

That being said, there are times where a less-structured discussion of
sketches and storyboards after heat map might make sense.

Opening this PR for discussion. If we want to keep this, should it make
its way onto one of the schedules?

If we don't want to keep there, where should it redirect to?