Move under the sprint-guide subdirectory
We will ultimately hosting this at design.thoughtbot.com/sprint-guide.
This moves under that subdirectory, which also opens up a place for the
rest of design.thoughtbot.com to live.

We will eventually merge the repositories, either design-dot will come
in here or the other way around.

I've opted to keep data and styles up at the root so they can be shared
between everything, but I don't feel strongly about that and I'm open to
feedback.